requirements.txt
# Core Data Processing
pyspark==3.4.1

dbt-core==1.4.6
dbt-bigquery==1.4.6

# Data Manipulation
pandas>=1.5.0
pyarrow>=7.0.0
numpy>=1.22.0

# Cloud Integration
google-cloud-storage>=2.0.0
google-cloud-bigquery>=2.0.0

# Airflow
apache-airflow==2.7.3
apache-airflow-providers-google==10.4.0
apache-airflow-providers-apache-spark==3.0.0

# Testing
pytest==7.4.4
pytest-mock==3.12.0

# Utility
python-dotenv==1.0.0
requests==2.31.0
